fix(lex,parse): use `u64` for parsed unsigned integers (#49)
* fix(lex): parse all unsigned integers into `u64`

* fix(lex): validate lexed integer fits into `u64`

* fix(ast): use `u64` unsigned integer types for more flexibility

* fix(compile,parse): parse into `u64` type

+++ src/lex/token.rs
@@ -30,7 +30,10 @@ pub enum Token<'src> {
     Comment(&'src str),
     Ident(&'src str),
     String(&'src str),
-    Uint(usize),
+    /// `Uint` is any unsigned integer token; stored as a `u64` to ensure cross-
+    /// platform compatibility and maximum flexibility. Downstream phases should
+    /// enforce stricter size limits as needed.
+    Uint(u64),
 }
 
 /* ------------------------------ Enum: Keyword ----------------------------- */
